Split regridding and standardisation into separate CLIs.  Need to change some "standardise" test data - all regrid tests pass.
